+++ VGA_core_M4/src/front_layer.h
@@ -2,4 +2,3 @@
 
 char ** UART_tokens();
 void UART_tokens_clear(char **array);
-void UART_control(char **array, uint8_t *perr);
