+++ VGA_core_M4/src/logic_layer.h
@@ -1,0 +1,1 @@
+void UART_control(char **array, uint8_t *perr);
